requirement.txt
Django==1.11.4
django-forms-bootstrap==3.1.0
django-markdown-deux==1.0.5
django-pagedown==0.1.3
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
pytz==2017.2
